Turret Game. initial commit.
The purpose of this game is to become familiar with the "processing" language
and the "box2D" physics engine. The Turret game is based on popular games like
"Tanks" and "Worms". Enemies appear in random positions on the screen, and must
be defeated by shooting them with the user-controlled turret.

Controls:
	LEFT / RIGHT : aim turret
	ENTER : fire bullet

Former-commit-id: c0b3cf71662b3d7972fff0cd6f02dc2eabbb98ea