Replaced TAB with 2 spaces for compatibility with Processing IDE.
Processing IDE does not behave well in the presence of TAB characters,
formatting becomes an issue which inhibits the ability to edit text. Typing a
literal TAB in the Processing IDE treated as 2 consecutive spaces.

Replaces all TAB characters will 2 spaces to work around this issue. In the
future, I will not introduce TAB chararcters.


Former-commit-id: 276badec94757efdb8c858200c014f3148be3028