Myo instance is now handled by a Singleton.
Each Myo() instance consumes a lot of resources (it registers to listen to a
bunch of events and stuff like that). To keep the game performant, it is
necessary to ensure that only a single instance of the Myo() is even
instantiated.

This is a hack. The singleton should be enclosed within a static class. But some
nuances with the way Processing works is making this challenging.


Former-commit-id: 5bfedc6f700ba9b4aa51e6babc5c3d0cf18e9f4b