Added an event-driven Calibrate menu.
Calibration of the Myo Armband can now be performed by entering the
'CalibrationMenu' game state. Depending on the outcome, the user will be
redirected to either the 'CalibrationSuccess' or 'CalibrationFailure' menu,
where they can select to either retry calibration or return to the MainMenu.

The calibration instructions currently appear in the terminal window. This will
be changed in a future commit so that users see instructions and a count down
timer within the main application window.

Calibration can fail for 2 reasons:

  1. No Armband is connected to the computer. (i.e. MyoNotConnectedException)
  2. An Armbarnd is connected, but no EMG signal is detected. (i.e., CalibrationFailedException)

In both cases, game control is reduced to the use of the keyboard.


Former-commit-id: d275dd4683d4e1e0742eb326398f5365129cd207