Fix. subscriptions to event listener set are performed prior to sending events.
In each frame, the set of objects listening for events was being consolidated
(applying additions and deletions to the set) only after the current queue of
events were fired. This meant that a newly subscribed listener would miss any
events occuring in the same frame as their subscription.

Now, in the EventManager,The set of event listeners is consolidated (additions
and deletions are applied) before any events are sent.


Former-commit-id: d760866b0403ef86f44e792e4d14324a75f51af0