Add sliders for io settings adjustments.
Import G4P GUI library for the io settings sliders. When the sliders are altered
the new values from the sliders are saved to data/xml_data/save_data.xml.
Add images for the slider gui. Add text components for the 'more' and 'less'
slider limits (the G4P library does not accept String for the slider limits).


Former-commit-id: bbb6b87d6b4dfa276d7f7256aa999a3911b8d03f