Options::GameRecord screen, table rows consistently appear in order.
Previously, the rows of the GameRecord table were not guaranteed to appear in
any consistent order. This was being caused because querying for the game
object using their tag did not necessarily retrieve them in the order that they
are displayed in the table.

Each row of the table now has a unique tag (ex., "stats_row_0", "stats_row_1"),
so that their order of their retrieval can be guaranteed.

closes #32.


Former-commit-id: c630367e426908186418b77a3bc097945a1edd16