Reduced the sticky platform coefficient of friction.
With the previous coefficient of friction settings, it was extremely difficult
to move across a long "sticky" platform when using the Myo-controls. The player
essentially had to resort to jumping across the platform to avoid losing the
game.

Reduced the coefficient of friction to make it possible for a player to
successfully move across a long "sticky" platform.


Former-commit-id: d901e2ae22d65326438d6febf19733313637a5ac