Obstacles no longer spawn over gaps.
Previously, it was possible for an obstacle to spawn in the middle of a gap
(seemingly in mid-air) making it impossible for the player to pass through the
gap. This was possible because the spawn position was being by randomly
selecting an offset from the middle of a platform. On extremely small platforms,
this offset could be large enough to cause the obstacle to spawn beyond the end of
the platform.

The minimum and maximum limits for the randomly selected obstacle spawn-position
now take the platform's width into consideration.

closes #13.


Former-commit-id: 2b8c7f7d445a87349bb159d8568d84bcaca5e900