Coins now only spawn on platforms currently below screen.
A known issue within the game is that spawning a coin on a platform causes the
animation of the platform to "stutter" momentarily. To work aroud this issue,
platforms are spawned well below the bottom of the screen, and coins are now
only spawned on platforms which have not yet risen high enough to be visible on
the screen.


Former-commit-id: 19152b5f61e5392fea4a28c37f1f190d84ef98d8