Adjusted platform spawn height to allow more coins to spawn.
Since coins can only spawn on platforms beneath the bottom of the visible
screen, it is often the case that when a coin-spawn is attempted, no suitable
platforms exist, so the spawn is simply aborted.

Lowered the platform spawn-height to allow a greater chance for a coin-spawn to
find a suitable platform.


Former-commit-id: dd19089145714c00eab13593d8359c46f109463f