Added "Quit Game" button to the pause/settings menu.
The player can now end an ongoing game from the pause-menu. All coins and score
collected so far will be collected (quiting a game is equivalent to getting a
"Game Over").

Since the pause screen is shared between the "Pause" and "System Settings"
screen, the addition of this button is somewhat of a hack. The quit button is
statically defined in the XML code, but is dynamically removed when the screen
is rendered as the settings menu. Additionally, the screen title is dynamically
changed from "System Settings" to "Paused" when the screen as rendered as the
pause screen.

closes #39.


Former-commit-id: 3bdcd2cb8327266027f7c48850138528c39d7550