fix. EmgManager::poll() now uses current EmgSamplingPolicy.
Previously, the EmgManager stored the EmgSamplingPolicy setting in an instance
variable. This meant that, no matter which was currently selected, the
EmgManager would use the policy selected when it was first instantiated.

EmgManager now checks the sampling policy during a poll() to ensure it is using
the most recent setting.

** while this bug has been present since the EmgManager was introduced, it was
being disguised because up until commit: e2181, the PlayerControllerComponent
was also correcting the raw input to correspond with the currently selected
sampling policy.


Former-commit-id: 50f0ebbb58272e2be9430dd12654eb472bf9fed3