Calibration snapshot is now triggered by pressing spacebar.
Previously, calibration involved holding a muscle contraction while the
calibration screen counts down and then takes a sensor snapshot. This process
was less than ideal because, depending on the user, the countdown may be too
long (for those with difficulty holding contractions), or too short (for those
unfamiliar with the process).

The calibration screens now present the player with an action to perform (i.e.,
flexion/extension) and prompts the user to hit SPACE when they are ready to
capture sensor readings. This allows the calibration to proceed at a reasonable
pace for all users.


Former-commit-id: 2c20bc3c0ff0c7bd07a2f4f205ea5ec243868533