Force Momo through gaps once he is close enough in Direction Assist mode.
Direction Assist mode was created to allow individuals with little-to-no muscle
control to still benefit from our game. However, in Direction Assist mode, it is
possible to 'skip' over gaps if full intensity muscle contractions are
performed.

In this fix, Momo is now forced to fall through gaps. This is accomplished by
teleporting him to the center of the gap and ignoring all input once he is
within a certain threshold distance from the gap. This ensures that the player
can always proceed through gameplay in Direction Assist mode, regardless of
their ability to control the strength/timing of their muscle contractions.


Former-commit-id: e6435c3134cabeb700a92a2fdc1469d38ecd4da6