Added a visual 'excessive force' warning.
In games, it is often instinctive to push buttons harder to try to get more from
the game (consider a racing game, where players often press the gas button as
hard as possible to try and go faster). We noticed the same behavior in The
Falling of Momo, and in our case, this is couter-productive to training goals
('trying harder' often leads to unintentional co-contraction).

Added a visual warning, in the form of a red circle around Momo, that warns
players when they are 'working harder' than necessary (i.e., contracting harder
than necessary for their current sensitivity settings). The width of the circle
grows as the contraction intensity increases further above what is necessary.

*** this warning should really be encapsulated into a proper 'gameObject', and
added and removed from the gameWorld as necessary.


Former-commit-id: c2a93b034e01c2d4f8f7053846c24ba1d7d537bc