Migrated from Sound to Minim.
The newest versions of the Sound library and Processing don't work nicely
together. To avoid further pain in the future, we've decided to switch to the
Minim audio library.

In this commit, I've created an abstract SoundManager class which encapsulates
all details of the specific sound library we decide to use. This means that in
the future, if Minim decides to change their API, or if we decide to switch
again, the effect on the rest of our Momo code should be minimal. All
sound-related calls in the code-base now delegate to the SoundManager.

Note: In this commit, I removed the sound-related add() and rate() calls, since
they were never called with anything except for default values.

Former-commit-id: d37e879321dc361945c1781e3bf3f328d8daab67