Fitts results saved to a single per-user file, flushed after each bonus level.
All Fitts results for a given user are now stored in a single file. Previously, a separate file was created for each "game" the user plays (i.e., every time the quit/game-over, a new file is created). Separate runs can be distinguished using the included timestamp on each row of data.


Former-commit-id: 5d5dd5d8b3b5c9b27592b4a7b2cc972216cd8ffa