Added throbber animation when waiting for armband to connect.
It takes a few moments for the game to establish a bluetooth connection with the myo armband. Since this is all happening on a single thread, the games hangs durign this time.

Set the cursor to a throbber during this hang to notify the user that something is going on in the background.

closes #77


Former-commit-id: b0c54157c354356329fd32cf794b108e255f72e1