Fix. Momo no longer gets stuck between platform/target on bonus level.
Made Momo a (physics) circle in the bonus level, which means he no longer has corners to get stuck in the crack between a platform and target. When Momo is a circle, he will sometimes "lauch" off of the target when passing over it with a high speed. I prevented this launch from occuring by setting a limit on the maximum upward speed he can attain in the bonus level.


Former-commit-id: c0fb6ec7b5c3e2b16cc29171f464c475b44ac214