Fix. Load existing bonus level logging data on game start.
In a previous commit, I modified the Fitts Law logging data so that bonus level data from all of a user's games were logged to a single output file. The intended behavior of this was that later results would simply be appended to the same file. However, this was not the case, and the logfile was being zero-ed at the beginning of each game.

A user's existing log data is now loaded upon game start, so that all subsequent bonus level results can be appended to the existing ones.

** note that I now have to use modulo arithmetic when indexing to find the gap positions.


Former-commit-id: c7bffef19823b6a6153946a35a31f579d4788595