Remove bonus-level "drifting".
To ensure accurate and consistent fitt's law configurations, Momo is explicitly positioned in the middle of the break-through platform once he "breaks through" (i.e., dwells for the required period or jumps while on the platform). Immediately following this "break through" left/right input control is disabled because Momo is no longer in contact with a game platform. However, it is possible for Momo to "drift" from this centered position if the player was moving left or right immediately before the "break through" occurred. This is because Momo's velocity is maintained, even though his position was explicitly updated.

To address this problem, Momo's velocity (x-component) is explicitly zero-ed every time a break platform is "broken through" in the bonus level. The player will be unable to provide left/right control input to the game until Momo comes into contact with another game platform, so Momo will remain stationary (in the X direction) for his entire fall.

TODO: this is not a very nice fix and deserves refactoring. Momo's rigid-body is explicitly updated in 3 places.

  1. rename the event fired when Momo "breaks through" a platform to be more meaningful
  2. add required parameter fields to this event to hold platforms position
  3. perform all explicit rigid-body updates in the handleEvents() method