Removed adjustable breakthrough platform dwell time.
Previously, the player could adjust the dwell time needed to breakthrough a
crumbling platform through the game settings menu. Since any adjustments to
this were actually undesirable (in practice, we always wanted to stick with the
2 second default), I removed the ability to adjust this.

Note: During gameplay, the dwell time paramater seemed to occasionally get
corrupted, and change from 2 seconds to a much longer dwell time. I believe
that this has to do with the fact that CounterComponent unconditionally
(Components.pde >> CounterComponent class) increments the dwelltime whenever
clicked, even though CounterComponents are also used in other places in the
game (e.g., in the Manual Calibration screen).  I haven't tested and confirmed
this as the cause of the bug.